@@ -1,14 +1,14 @@
-const fs = require('fs');
-const { HttpClient } = require('@actions/http-client');
+import fs from 'fs';
+import { HttpClient } from '@actions/http-client';
 
 /**
  * Downloads a file using `HttpClient` class. Supports HTTP redirection out-of-box.
  *
- * @param {String} url Source URL of file to download.
- * @param {String} dest Destination path where to save the downloaded file.
- * @returns {Promise}
+ * @param {string} url Source URL of file to download.
+ * @param {string} dest Destination path where to save the downloaded file.
+ * @returns {Promise<true>}
  */
-module.exports = (url, dest) => new Promise((resolve, reject) => {
+const downloadFile = (url: string, dest: string): Promise<true> => new Promise((resolve, reject) => {
     return new HttpClient().get(url)
         .then((res) => {
             const file = fs.createWriteStream(dest);
@@ -17,7 +17,7 @@ module.exports = (url, dest) => new Promise((resolve, reject) => {
                 .pipe(file)
                 .on('finish', () => {
                     file.close(() => {
-                        resolve();
+                        resolve(true);
                     });
                 }).on('error', (error) => {
                     fs.unlink(dest, () => {
@@ -28,3 +28,5 @@ module.exports = (url, dest) => new Promise((resolve, reject) => {
             reject(error.message);
         });
 });
+
+export default downloadFile;
