@@ -1,11 +1,12 @@
-const process = require('process');
-const fs = require('fs');
-const path = require('path');
-const core = require('@actions/core');
-const exec = require('@actions/exec');
-const { mkdirP } = require('@actions/io');
+import process from 'process';
+import fs from 'fs';
+import path from 'path';
+import * as core from '@actions/core';
+import * as exec from '@actions/exec';
+import { mkdirP } from '@actions/io';
 
-const buildPackage = require('../src/build-package');
+import buildPackage from '../src/build-package';
+import { EnvironmentVariables } from '../src/types/env-functions';
 
 jest.mock('@actions/core');
 jest.mock('@actions/exec');
@@ -27,9 +28,11 @@ const codeCoverage = true;
 const os = 'ubuntu-20.04';
 const macOs = 'macos-10.15';
 const compiler = 'gnu-10';
+const errorObject = new Error('spawn /bin/sh ENOENT');
+const emptyObject = {};
 
 // Base environment object, we will take care not to modify it.
-const env = {
+const env: EnvironmentVariables = {
     CC: 'gcc-10',
     CXX: 'g++-10',
     FC: 'gfortran-10',
@@ -44,6 +47,8 @@ describe('buildPackage', () => {
             ...env,
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(true);
@@ -56,13 +61,11 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation(() => Promise.resolve(1));
+        (exec.exec as jest.Mock).mockResolvedValue(1);
 
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('supports cmake switch', async () => {
@@ -72,7 +75,9 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        let isBuilt = await buildPackage(repository, sourceDir, installDir, !cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
+        const isBuilt = await buildPackage(repository, sourceDir, installDir, !cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(true);
         expect(core.info).toHaveBeenCalledWith('==> configurePath: cmake');
@@ -86,13 +91,15 @@ describe('buildPackage', () => {
             ...env,
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         let isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, !codeCoverage, os, compiler, testEnv1);
 
         expect(isBuilt).toBe(true);
         expect(core.info).toHaveBeenCalledWith('==> configurePath: ecbuild');
         expect(core.info).toHaveBeenCalledWith(`==> configureOptions: --prefix=${installDir}`);
 
-        core.info.mockReset();
+        (core.info as jest.Mock).mockReset();
 
         const testEnv2 = {
             ...env,
@@ -122,13 +129,15 @@ describe('buildPackage', () => {
             ...env,
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         let isBuilt = await buildPackage(repository, sourceDir, installDir, !cmake, cmakeOptions, ctestOptions, test, !codeCoverage, os, compiler, testEnv1);
 
         expect(isBuilt).toBe(true);
         expect(core.info).toHaveBeenCalledWith('==> configurePath: cmake');
         expect(core.info).toHaveBeenCalledWith(`==> configureOptions: -DCMAKE_INSTALL_PREFIX=${installDir},${testCmakeOptions}`);
 
-        core.info.mockReset();
+        (core.info as jest.Mock).mockReset();
 
         const testEnv2 = {
             ...env,
@@ -157,12 +166,14 @@ describe('buildPackage', () => {
             ...env,
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         let isBuilt = await buildPackage(repository, sourceDir, installDir, !cmake, cmakeOptions, ctestOptions, test, !codeCoverage, os, compiler, testEnv1);
 
         expect(isBuilt).toBe(true);
         expect(core.info).toHaveBeenCalledWith(`==> testOptions: ${testCtestOptions}`);
 
-        core.info.mockReset();
+        (core.info as jest.Mock).mockReset();
 
         const testEnv2 = {
             ...env,
@@ -180,6 +191,8 @@ describe('buildPackage', () => {
         const testEnv = {
             ...env,
         };
+
+        (exec.exec as jest.Mock).mockResolvedValue(0);
 
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
@@ -205,14 +218,18 @@ describe('buildPackage', () => {
         const cmakeOptionsFile = path.join(sourceDir, '.github', '.cmake-options');
         const cmakeOptionsFileContent = testCmakeOptions.join(' ');
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         const existsSync = jest.spyOn(fs, 'existsSync');
         existsSync.mockImplementation((path) => {
             if (path === cmakeOptionsFile) return true;
+            return false;
         });
 
         const readFileSync = jest.spyOn(fs, 'readFileSync');
         readFileSync.mockImplementation((path) => {
             if (path === cmakeOptionsFile) return cmakeOptionsFileContent;
+            return '';
         });
 
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, !codeCoverage, os, compiler, testEnv);
@@ -220,9 +237,6 @@ describe('buildPackage', () => {
         expect(isBuilt).toBe(true);
         expect(core.info).toHaveBeenCalledWith(`==> Found ${cmakeOptionsFile}: ${cmakeOptionsFileContent}`);
         expect(core.info).toHaveBeenCalledWith(`==> configureOptions: --prefix=${installDir},${testCmakeOptions}`);
-
-        existsSync.mockReset();
-        readFileSync.mockReset();
     });
 
     it('supports backwards compatible magic path for cmake options', async () => {
@@ -241,6 +255,8 @@ describe('buildPackage', () => {
         const deprecatedCmakeOptionsFile = path.join(sourceDir, '.github', '.compiler-flags');
         const deprecatedCmakeOptionsFileContent = testDeprecatedCmakeOptions.join(' ');
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         const existsSync = jest.spyOn(fs, 'existsSync');
         existsSync.mockImplementation((path) => {
             if (path === deprecatedCmakeOptionsFile) return true;
@@ -250,6 +266,7 @@ describe('buildPackage', () => {
         const readFileSync = jest.spyOn(fs, 'readFileSync');
         readFileSync.mockImplementation((path) => {
             if (path === deprecatedCmakeOptionsFile) return deprecatedCmakeOptionsFileContent;
+            return '';
         });
 
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, !codeCoverage, os, compiler, testEnv);
@@ -258,9 +275,6 @@ describe('buildPackage', () => {
         expect(core.info).toHaveBeenCalledWith(`==> Found ${deprecatedCmakeOptionsFile}: ${deprecatedCmakeOptionsFileContent}`);
         expect(core.warning).toHaveBeenCalledWith('Magic file path `.github/.compiler-flags` has been deprecated, please migrate to `.github/.cmake-options`');
         expect(core.info).toHaveBeenCalledWith(`==> configureOptions: --prefix=${installDir},${testDeprecatedCmakeOptions}`);
-
-        existsSync.mockReset();
-        readFileSync.mockReset();
     });
 
     it('expands shell variables in all arguments', async () => {
@@ -312,7 +326,9 @@ describe('buildPackage', () => {
             testEnv1.VAR2,
         ];
 
-        let isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, !codeCoverage, os, compiler, testEnv1);
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
+        const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, !codeCoverage, os, compiler, testEnv1);
 
         expect(isBuilt).toBe(true);
         expect(core.info).toHaveBeenCalledWith(`==> configureOptions: --prefix=${installDir},${testCmakeOptions}`);
@@ -327,7 +343,7 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation((command, args) => {
+        (exec.exec as jest.Mock).mockImplementation((command, args) => {
             if (
                 command === 'sudo'
                 && args[0] === 'apt-get'
@@ -345,8 +361,6 @@ describe('buildPackage', () => {
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('adds code coverage compiler flags on supported platform', async () => {
@@ -355,6 +369,8 @@ describe('buildPackage', () => {
         const testEnv = {
             ...env,
         };
+
+        (exec.exec as jest.Mock).mockResolvedValue(0);
 
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
@@ -370,6 +386,8 @@ describe('buildPackage', () => {
             ...env,
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, macOs, compiler, testEnv);
 
         expect(isBuilt).toBe(true);
@@ -383,11 +401,13 @@ describe('buildPackage', () => {
             ...env,
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(true);
 
-        core.info.mock.calls.forEach((call) => {
+        (core.info as jest.Mock).mock.calls.forEach((call) => {
             const arg = call[0];
             if (!/^==> options\.env:/.test(arg)) return;
 
@@ -400,7 +420,7 @@ describe('buildPackage', () => {
     it('allows overriding of default options via current environment variables', async () => {
         expect.assertions(12);
 
-        const testEnv = {
+        const testEnv: EnvironmentVariables = {
             ...env,
             'CTEST_OUTPUT_ON_FAILURE': '0',
             'CMAKE_BUILD_PARALLEL_LEVEL': '1',
@@ -415,6 +435,8 @@ describe('buildPackage', () => {
             },
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(true);
@@ -427,7 +449,7 @@ describe('buildPackage', () => {
         expect(exec.exec).toHaveBeenCalledWith('env', ['genhtml', coverageFile, '--output-directory', coverageDir], options);
         expect(exec.exec).toHaveBeenCalledWith('env', ['cmake', '--install', '.'], options);
 
-        core.info.mock.calls.forEach((call) => {
+        (core.info as jest.Mock).mock.calls.forEach((call) => {
             const arg = call[0];
             if (!/^==> options\.env:/.test(arg)) return;
 
@@ -443,6 +465,8 @@ describe('buildPackage', () => {
         const testEnv = {
             ...env,
         };
+
+        (exec.exec as jest.Mock).mockResolvedValue(0);
 
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
@@ -467,6 +491,8 @@ describe('buildPackage', () => {
                 ...testEnv,
             },
         };
+
+        (exec.exec as jest.Mock).mockResolvedValue(0);
 
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
@@ -499,6 +525,8 @@ describe('buildPackage', () => {
             },
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, !codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(true);
@@ -528,6 +556,8 @@ describe('buildPackage', () => {
             },
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, !test, !codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(true);
@@ -548,7 +578,7 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation((command, args) => {
+        (exec.exec as jest.Mock).mockImplementation((command, args) => {
             if (
                 command === 'env'
                 && args[0] === 'ecbuild'
@@ -564,8 +594,6 @@ describe('buildPackage', () => {
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, !test, !codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('returns false if build command failed', async () => {
@@ -575,7 +603,7 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation((command, args) => {
+        (exec.exec as jest.Mock).mockImplementation((command, args) => {
             if (
                 command === 'env'
                 && args[0] === 'cmake'
@@ -591,8 +619,6 @@ describe('buildPackage', () => {
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, !test, !codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('returns false if test command failed', async () => {
@@ -602,7 +628,7 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation((command, args) => {
+        (exec.exec as jest.Mock).mockImplementation((command, args) => {
             if (
                 command === 'env'
                 && args[0] === 'ctest'
@@ -616,8 +642,6 @@ describe('buildPackage', () => {
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('returns false if code coverage collection command failed', async () => {
@@ -627,7 +651,7 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation((command, args) => {
+        (exec.exec as jest.Mock).mockImplementation((command, args) => {
             if (
                 command === 'env'
                 && args[0] === 'lcov'
@@ -646,8 +670,6 @@ describe('buildPackage', () => {
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('returns false if code coverage cleanup command failed', async () => {
@@ -657,7 +679,7 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation((command, args) => {
+        (exec.exec as jest.Mock).mockImplementation((command, args) => {
             if (
                 command === 'env'
                 && args[0] === 'lcov'
@@ -678,8 +700,6 @@ describe('buildPackage', () => {
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('returns false if code coverage report listing command failed', async () => {
@@ -689,7 +709,7 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation((command, args) => {
+        (exec.exec as jest.Mock).mockImplementation((command, args) => {
             if (
                 command === 'env'
                 && args[0] === 'lcov'
@@ -705,8 +725,6 @@ describe('buildPackage', () => {
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('returns false if code coverage report generation command failed', async () => {
@@ -716,7 +734,7 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation((command, args) => {
+        (exec.exec as jest.Mock).mockImplementation((command, args) => {
             if (
                 command === 'env'
                 && args[0] === 'genhtml'
@@ -733,8 +751,6 @@ describe('buildPackage', () => {
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('returns false if install command failed', async () => {
@@ -744,7 +760,7 @@ describe('buildPackage', () => {
             ...env,
         };
 
-        exec.exec.mockImplementation((command, args) => {
+        (exec.exec as jest.Mock).mockImplementation((command, args) => {
             if (
                 command === 'env'
                 && args[0] === 'cmake'
@@ -760,8 +776,6 @@ describe('buildPackage', () => {
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 
     it('extends environment object with install paths', async () => {
@@ -782,6 +796,8 @@ describe('buildPackage', () => {
             [`${repo.toUpperCase()}_DIR`]: installDir,
             [`${repo.toUpperCase()}_PATH`]: installDir,
         };
+
+        (exec.exec as jest.Mock).mockResolvedValue(0);
 
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, !codeCoverage, os, compiler, testEnv);
 
@@ -812,27 +828,31 @@ describe('buildPackage', () => {
             COVERAGE_DIR: coverageDir,
         };
 
+        (exec.exec as jest.Mock).mockResolvedValue(0);
+
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(true);
         expect(testEnv).toStrictEqual(expectedEnv);
     });
 
-    it('returns false if command throws an error', async () => {
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('returns false if command throws an error ($error)', async ({ error }) => {
         expect.assertions(1);
 
         const testEnv = {
             ...env,
         };
 
-        exec.exec.mockImplementation(() => {
-            throw Error('spawn /bin/sh ENOENT');
+        (exec.exec as jest.Mock).mockImplementationOnce(() => {
+            throw error;
         });
 
         const isBuilt = await buildPackage(repository, sourceDir, installDir, cmake, cmakeOptions, ctestOptions, test, codeCoverage, os, compiler, testEnv);
 
         expect(isBuilt).toBe(false);
-
-        exec.exec.mockReset();
     });
 });
