@@ -1,10 +1,12 @@
-const cache = require('@actions/cache');
-const core = require('@actions/core');
-const { Octokit } = require('@octokit/core');
-const { getCacheKey, restoreCache, saveCache } = require('../src/cache-functions');
-const crypto = require('crypto');
-const fastFolderSize = require('fast-folder-size');
-const { version } = require('../package.json');
+import * as cache from '@actions/cache';
+import * as core from '@actions/core';
+import { Octokit } from '@octokit/core';
+import crypto from 'crypto';
+import fastFolderSize from 'fast-folder-size';
+
+import { getCacheKey, restoreCache, saveCache } from '../src/cache-functions';
+import { version } from '../package.json';
+import { EnvironmentVariables } from '../src/types/env-functions';
 
 jest.mock('@actions/core');
 jest.mock('@actions/cache');
@@ -21,6 +23,8 @@ const compiler = 'gnu-10';
 const cacheSuffix = '';
 const installDir = '/path/to/install/repo';
 const sha = 'f0b00fd201c7ddf14e1572a10d5fb4577c4bd6a2';
+const errorObject = new Error('Oops!');
+const emptyObject = {};
 
 const env = {
     CMAKE_VERSION: '3.20.5',
@@ -39,7 +43,7 @@ const resolveHeadSha = () => Promise.resolve({
     },
 });
 
-let cacheKey;
+let cacheKey: string;
 
 describe('getCacheKey', () => {
     it('returns a consistent cache key', async () => {
@@ -58,7 +62,7 @@ describe('getCacheKey', () => {
 
         const cacheKeySha = crypto.createHash('sha1').update(cacheKeyStr).digest('hex');
 
-        Octokit.prototype.constructor.mockImplementation((options) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation((options) => {
             if (!options.auth) throw Error(`Octokit authentication missing, did you pass the auth key?`);
 
             return {
@@ -77,8 +81,6 @@ describe('getCacheKey', () => {
 
         expect(core.info).toHaveBeenCalledWith(`==> Branch: ${branch}`);
         expect(core.info).toHaveBeenCalledWith(`==> Ref: heads/${branch}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
     it('supports tags', async () => {
@@ -100,7 +102,7 @@ describe('getCacheKey', () => {
 
         const cacheKeySha = crypto.createHash('sha1').update(cacheKeyStr).digest('hex');
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
@@ -109,8 +111,6 @@ describe('getCacheKey', () => {
         expect(cacheKey).toStrictEqual(`${os}-${compiler}-${repo}-${cacheKeySha}`);
         expect(core.info).toHaveBeenCalledWith(`==> Branch: ${testTag}`);
         expect(core.info).toHaveBeenCalledWith(`==> Ref: tags/${testTag}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
     it('skips current repository as a dependency', async () => {
@@ -134,21 +134,19 @@ describe('getCacheKey', () => {
 
         const cacheKeySha = crypto.createHash('sha1').update(cacheKeyStr).digest('hex');
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
         const { cacheKey } = await getCacheKey(repository, branch, githubToken, os, compiler, cacheSuffix, testEnv);
 
         expect(cacheKey).toStrictEqual(`${os}-${compiler}-${repo}-${cacheKeySha}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
     it('returns cache key if dependencies object is undefined', async () => {
         expect.assertions(1);
 
-        const testEnv = {
+        const testEnv: EnvironmentVariables = {
             ...env,
         };
 
@@ -157,15 +155,13 @@ describe('getCacheKey', () => {
         const cacheKeyStr = `v=${version}${cacheSuffix}::cmake=${testEnv.CMAKE_VERSION}::${repo}=${sha}`;
         const cacheKeySha = crypto.createHash('sha1').update(cacheKeyStr).digest('hex');
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
         const { cacheKey } = await getCacheKey(repository, branch, githubToken, os, compiler, cacheSuffix, testEnv);
 
         expect(cacheKey).toStrictEqual(`${os}-${compiler}-${repo}-${cacheKeySha}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
     it('returns cache key if dependencies object is empty', async () => {
@@ -179,19 +175,21 @@ describe('getCacheKey', () => {
         const cacheKeyStr = `v=${version}${cacheSuffix}::cmake=${testEnv.CMAKE_VERSION}::${repo}=${sha}`;
         const cacheKeySha = crypto.createHash('sha1').update(cacheKeyStr).digest('hex');
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
         const { cacheKey } = await getCacheKey(repository, branch, githubToken, os, compiler, cacheSuffix, testEnv);
 
         expect(cacheKey).toStrictEqual(`${os}-${compiler}-${repo}-${cacheKeySha}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
-    it('logs error if repository HEAD fetch fails', async () => {
-        expect.assertions(3);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('logs error if repository HEAD fetch fails ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
         const testEnv = {
             ...env,
@@ -206,11 +204,9 @@ describe('getCacheKey', () => {
 
         const cacheKeySha = crypto.createHash('sha1').update(cacheKeyStr).digest('hex');
 
-        const errorMessage = 'Oops!';
-
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: () => {
-                throw new Error(errorMessage);
+                throw error;
             },
         }));
 
@@ -218,9 +214,9 @@ describe('getCacheKey', () => {
 
         expect(cacheKey).toStrictEqual(`${os}-${compiler}-${repo}-${cacheKeySha}`);
         expect(core.info).toHaveBeenCalledWith(`==> result.headSha: undefined`);
-        expect(core.warning).toHaveBeenCalledWith(`Error getting repository HEAD for ${repo}: ${errorMessage}`)
 
-        Octokit.prototype.constructor.mockReset();
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error getting repository HEAD for ${repo}: ${error.message}`)
     });
 
     it('invalidates the cache if suffix is supplied', async () => {
@@ -239,7 +235,7 @@ describe('getCacheKey', () => {
 
         const cacheKeySha = crypto.createHash('sha1').update(cacheKeyStr).digest('hex');
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
@@ -264,8 +260,6 @@ describe('getCacheKey', () => {
         expect(newResult.cacheKey).toStrictEqual(`${os}-${compiler}-${repo}-${newCacheKeySha}`);
         expect(newResult.cacheKey).not.toStrictEqual(result.cacheKey);
         expect(newResult.headSha).toStrictEqual(result.headSha);
-
-        Octokit.prototype.constructor.mockReset();
     });
 });
 
@@ -273,41 +267,39 @@ describe('restoreCache', () => {
     it('restores package from cache if found', async () => {
         expect.assertions(4);
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
         for (const mockCacheHit of [false, true]) {
-            cache.restoreCache.mockResolvedValue(mockCacheHit);
+            (cache.restoreCache as jest.Mock).mockResolvedValueOnce(mockCacheHit);
 
             const cacheHit = await restoreCache(repository, branch, githubToken, installDir, os, compiler, cacheSuffix, env);
 
             expect(cacheHit).toBe(mockCacheHit);
             expect(cache.restoreCache).toHaveBeenCalledWith([installDir], cacheKey);
         }
-
-        Octokit.prototype.constructor.mockReset();
-        cache.restoreCache.mockReset();
     });
 
-    it('catches unexpected restore cache errors', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('catches unexpected restore cache errors ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
-        const errorMessage = 'Oops!';
-
-        cache.restoreCache.mockRejectedValue(new Error(errorMessage));
+        (cache.restoreCache as jest.Mock).mockRejectedValueOnce(error);
 
         const cacheHit = await restoreCache(repository, branch, githubToken, installDir, os, compiler, cacheSuffix, env);
 
         expect(cacheHit).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error restoring cache for ${repository}: ${errorMessage}`);
 
-        Octokit.prototype.constructor.mockReset();
-        cache.restoreCache.mockReset();
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error restoring cache for ${repository}: ${error.message}`);
     });
 });
 
@@ -315,36 +307,32 @@ describe('saveCache', () => {
     it('saves package to cache', async () => {
         expect.assertions(4);
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
             request: resolveHeadSha,
         }));
 
-        fastFolderSize.mockImplementation((f, cb) => {
+        (fastFolderSize as jest.Mock).mockImplementation((f, cb) => {
             if (f) cb(null, 1024);
         });
 
         for (const mockIsSaved of [false, true]) {
-            cache.saveCache.mockResolvedValue(mockIsSaved);
+            (cache.saveCache as jest.Mock).mockResolvedValueOnce(mockIsSaved);
 
             const isSaved = await saveCache(repository, branch, githubToken, installDir, os, compiler, cacheSuffix, env);
 
             expect(isSaved).toBe(mockIsSaved);
             expect(cache.saveCache).toHaveBeenCalledWith([installDir], cacheKey);
         }
-
-        Octokit.prototype.constructor.mockReset();
-        fastFolderSize.mockReset();
-        cache.saveCache.mockReset();
     });
 
     it('does not save empty package to cache', async () => {
         expect.assertions(2);
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
-        fastFolderSize.mockImplementation((f, cb) => {
+        (fastFolderSize as jest.Mock).mockImplementationOnce((f, cb) => {
             if (f) cb(null, 0);
         });
 
@@ -352,58 +340,53 @@ describe('saveCache', () => {
 
         expect(isSaved).toBe(false);
         expect(cache.saveCache).not.toHaveBeenCalled();
-
-        Octokit.prototype.constructor.mockReset();
-        fastFolderSize.mockReset();
     });
 
     it('does not save cache on key collisions', async () => {
         expect.assertions(2);
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
-        fastFolderSize.mockImplementation((f, cb) => {
+        (fastFolderSize as jest.Mock).mockImplementationOnce((f, cb) => {
             if (f) cb(null, 1024);
         });
 
         const errorMessage = `Unable to reserve cache with key ${cacheKey}, another job may be creating this cache.`;
 
-        cache.saveCache.mockRejectedValue(new Error(errorMessage));
+        (cache.saveCache as jest.Mock).mockRejectedValueOnce(new Error(errorMessage));
 
         const isSaved = await saveCache(repository, branch, githubToken, installDir, os, compiler, cacheSuffix, env);
 
         expect(isSaved).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`Error saving cache for ${repository}: ${errorMessage}`);
-
-        Octokit.prototype.constructor.mockReset();
-        fastFolderSize.mockReset();
-        cache.saveCache.mockReset();
     });
 
-    it('catches unexpected save cache errors', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('catches unexpected save cache errors ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
+        (Octokit.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             request: resolveHeadSha,
         }));
 
-        fastFolderSize.mockImplementation((f, cb) => {
+        (fastFolderSize as jest.Mock).mockImplementationOnce((f, cb) => {
             if (f) cb(null, 1024);
         });
 
-        const errorMessage = 'Oops!';
-
-        cache.saveCache.mockRejectedValue(new Error(errorMessage));
+        (cache.saveCache as jest.Mock).mockImplementationOnce(() => {
+            throw error;
+        });
 
         const isSaved = await saveCache(repository, branch, githubToken, installDir, os, compiler, cacheSuffix, env);
 
         expect(isSaved).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error saving cache for ${repository}: ${errorMessage}`);
 
-        Octokit.prototype.constructor.mockReset();
-        fastFolderSize.mockReset();
-        cache.saveCache.mockReset();
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error saving cache for ${repository}: ${error.message}`);
     });
 });
