@@ -1,13 +1,13 @@
-const process = require('process');
-const fs = require('fs');
-const { Buffer } = require('buffer');
-const core = require('@actions/core');
-const { Octokit } = require('@octokit/core');
-const AdmZip = require('adm-zip');
-const filesize = require('filesize');
-const tar = require('tar');
+import process from 'process';
+import fs from 'fs';
+import { Buffer } from 'buffer';
+import * as core from '@actions/core';
+import { Octokit } from '@octokit/core';
+import AdmZip from 'adm-zip';
+import filesize from 'filesize';
+import tar from 'tar';
 
-const downloadArtifact = require('../src/download-artifact');
+import downloadArtifact from '../src/download-artifact';
 
 jest.mock('@actions/core');
 jest.mock('@actions/io');
@@ -33,7 +33,8 @@ const artifactPath = `${downloadDir}/${artifactName}`;
 const tarPath = `${artifactPath}/${artifactName}.tar`;
 const dependenciesPath = `${artifactPath}/${artifactName}-dependencies.json`;
 const errorStatusCode = 500;
-const errorMessage = 'Oops!';
+const errorObject = new Error('Oops!');
+const emptyObject = {};
 
 const resolveWorkflowRuns = () => Promise.resolve({
     status: 200,
@@ -49,7 +50,7 @@ const resolveWorkflowRuns = () => Promise.resolve({
     },
 });
 
-const resolveWorkflowRunArtifacts = (targetArtifactName) => Promise.resolve({
+const resolveWorkflowRunArtifacts = (targetArtifactName: string): Promise<Record<string, unknown>> => Promise.resolve({
     status: 200,
     data: {
         artifacts: [
@@ -131,11 +132,11 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation((options) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation((options) => {
             if (!options.auth) throw Error(`Octokit authentication missing, did you pass the auth key?`);
 
             return {
-                request: (route) => {
+                request: (route: string) => {
                     switch (route) {
                     case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                         return resolveWorkflowRuns();
@@ -150,18 +151,19 @@ describe('downloadArtifact', () => {
             };
         });
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
         const existsSync = jest.spyOn(fs, 'existsSync');
-        existsSync.mockImplementation((path) => {
+        existsSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return false;
+            return true;
         });
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => true);
+        unlinkSync.mockImplementationOnce(() => true);
 
         const isArtifactDownloaded = await downloadArtifact(repository, branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
@@ -187,10 +189,6 @@ describe('downloadArtifact', () => {
 
             expect(core.info).toHaveBeenCalledWith(`  ${action}: ${filepath}`);
         });
-
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('returns true if dependencies match', async () => {
@@ -209,8 +207,8 @@ describe('downloadArtifact', () => {
             },
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -224,22 +222,24 @@ describe('downloadArtifact', () => {
             },
         }));
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
         const existsSync = jest.spyOn(fs, 'existsSync');
-        existsSync.mockImplementation((path) => {
+        existsSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return true;
+            return false;
         });
 
         const readFileSync = jest.spyOn(fs, 'readFileSync');
-        readFileSync.mockImplementation((path) => {
+        readFileSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return JSON.stringify({
                 [dependency1]: dependency1Sha,
                 [dependency2]: dependency2Sha,
             });
+            return '';
         });
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
@@ -249,12 +249,6 @@ describe('downloadArtifact', () => {
 
         expect(isArtifactDownloaded).toBe(true);
         expect(core.info).toHaveBeenCalledWith(`==> Found ${dependenciesPath}`);
-
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        existsSync.mockReset();
-        readFileSync.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('looks for differently named artifact in case of ecbuild', async () => {
@@ -266,8 +260,8 @@ describe('downloadArtifact', () => {
 
         const ecbuildArtifactName = `ecbuild-${os}-cmake-${testEnv.CMAKE_VERSION}-${headSha}`;
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -281,22 +275,18 @@ describe('downloadArtifact', () => {
             },
         }));
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => true);
+        unlinkSync.mockImplementationOnce(() => true);
 
         const isArtifactDownloaded = await downloadArtifact('ecmwf/ecbuild', branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(true);
         expect(core.info).toHaveBeenCalledWith(`==> artifactName: ${ecbuildArtifactName}`);
-
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('returns false if no completed workflow runs were found', async () => {
@@ -306,8 +296,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return Promise.resolve({
@@ -326,22 +316,18 @@ describe('downloadArtifact', () => {
             },
         }));
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => true);
+        unlinkSync.mockImplementationOnce(() => true);
 
         const isArtifactDownloaded = await downloadArtifact('ecmwf/ecbuild', branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith('No completed successful workflow runs found for ecbuild');
-
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('returns false if no successful workflow runs were found', async () => {
@@ -351,8 +337,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return Promise.resolve({
@@ -371,22 +357,18 @@ describe('downloadArtifact', () => {
             },
         }));
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => true);
+        unlinkSync.mockImplementationOnce(() => true);
 
         const isArtifactDownloaded = await downloadArtifact('ecmwf/ecbuild', branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith('No completed successful workflow runs found for ecbuild');
-
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('returns false if request for workflow runs errors out', async () => {
@@ -396,8 +378,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return Promise.resolve({
@@ -411,8 +393,6 @@ describe('downloadArtifact', () => {
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`Wrong response code while fetching workflow runs for ${repo}: ${errorStatusCode}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
     it('returns false if no workflow runs were found', async () => {
@@ -422,8 +402,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return Promise.resolve({
@@ -440,22 +420,24 @@ describe('downloadArtifact', () => {
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`No workflow runs found for ${repo}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
-    it('returns false if request for workflow runs fails', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('returns false if request for workflow runs fails ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
         const testEnv = {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
-                    throw Error(errorMessage);
+                    throw error;
                 }
             },
         }));
@@ -463,9 +445,9 @@ describe('downloadArtifact', () => {
         const isArtifactDownloaded = await downloadArtifact(repository, branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error fetching workflow runs for ${repo}: ${errorMessage}`);
 
-        Octokit.prototype.constructor.mockReset();
+        if (error) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error fetching workflow runs for ${repo}: ${error.message}`);
     });
 
     it('returns false if request for workflow artifacts errors out', async () => {
@@ -475,8 +457,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -492,24 +474,26 @@ describe('downloadArtifact', () => {
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`Wrong response code while fetching workflow run artifacts for ${repo}: ${errorStatusCode}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
-    it('returns false if request for workflow artifacts fails', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('returns false if request for workflow artifacts fails ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
         const testEnv = {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
                 case 'GET /repos/{owner}/{repo}/actions/runs/{run_id}/artifacts':
-                    throw Error(errorMessage);
+                    throw error;
                 }
             },
         }));
@@ -517,9 +501,9 @@ describe('downloadArtifact', () => {
         const isArtifactDownloaded = await downloadArtifact(repository, branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error fetching workflow run artifacts for ${repo}: ${errorMessage}`);
 
-        Octokit.prototype.constructor.mockReset();
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error fetching workflow run artifacts for ${repo}: ${error.message}`);
     });
 
     it('returns false if no workflow artifacts are found', async () => {
@@ -529,8 +513,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -549,8 +533,6 @@ describe('downloadArtifact', () => {
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`No workflow artifacts found for ${repo}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
     it('returns false if repository HEAD state does not match', async () => {
@@ -562,8 +544,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -584,22 +566,18 @@ describe('downloadArtifact', () => {
             },
         }));
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => true);
+        unlinkSync.mockImplementationOnce(() => true);
 
         const isArtifactDownloaded = await downloadArtifact(repository, branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`No suitable artifact found: ${repo}-${os}-${compiler}-${newSha}`);
-
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('returns false if no artifacts with expected name were found', async () => {
@@ -609,8 +587,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -622,22 +600,18 @@ describe('downloadArtifact', () => {
             },
         }));
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => true);
+        unlinkSync.mockImplementationOnce(() => true);
 
         const isArtifactDownloaded = await downloadArtifact('ecmwf/ecbuild', branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`No suitable artifact found: ecbuild-${os}-cmake-${testEnv.CMAKE_VERSION}-${headSha}`);
-
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('returns false if request for repository HEAD runs errors out', async () => {
@@ -647,8 +621,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -666,26 +640,29 @@ describe('downloadArtifact', () => {
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`Wrong response code while fetching repository HEAD for ${repo}: ${errorStatusCode}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
-    it('returns false if request for repository HEAD runs fails', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('returns false if request for repository HEAD fails ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
         const testEnv = {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
                 case 'GET /repos/{owner}/{repo}/actions/runs/{run_id}/artifacts':
                     return resolveWorkflowRunArtifacts(artifactName);
                 case 'GET /repos/{owner}/{repo}/git/ref/{ref}':
-                    throw Error(errorMessage);
+                    // eslint-disable-next-line jest/no-if
+                    throw error;
                 }
             },
         }));
@@ -693,9 +670,9 @@ describe('downloadArtifact', () => {
         const isArtifactDownloaded = await downloadArtifact(repository, branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error getting repository HEAD for ${repo}: ${errorMessage}`);
 
-        Octokit.prototype.constructor.mockReset();
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error getting repository HEAD for ${repo}: ${error.message}`);
     });
 
     it('returns false if request for downloading workflow run artifact errors out', async () => {
@@ -705,8 +682,8 @@ describe('downloadArtifact', () => {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -726,19 +703,21 @@ describe('downloadArtifact', () => {
 
         expect(isArtifactDownloaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`Wrong response code while downloading workflow run artifact for ${repo}: ${errorStatusCode}`);
-
-        Octokit.prototype.constructor.mockReset();
     });
 
-    it('returns false if request for downloading workflow run artifact fails', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('returns false if request for downloading workflow run artifact fails ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
         const testEnv = {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -747,7 +726,7 @@ describe('downloadArtifact', () => {
                 case 'GET /repos/{owner}/{repo}/git/ref/{ref}':
                     return resolveHeadSha();
                 case 'GET /repos/{owner}/{repo}/actions/artifacts/{artifact_id}/{archive_format}':
-                    throw Error(errorMessage);
+                    throw error;
                 }
             },
         }));
@@ -755,20 +734,24 @@ describe('downloadArtifact', () => {
         const isArtifactDownloaded = await downloadArtifact(repository, branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error downloading workflow run artifact for ${repo}: ${errorMessage}`);
 
-        Octokit.prototype.constructor.mockReset();
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error downloading workflow run artifact for ${repo}: ${error.message}`);
     });
 
-    it('returns false if extracting artifact TAR fails', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('returns false if extracting artifact TAR fails ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
         const testEnv = {
             ...env,
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -782,23 +765,21 @@ describe('downloadArtifact', () => {
             },
         }));
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
-        tar.x.mockImplementation(() => {
-            throw new Error(errorMessage);
+        (tar.x as jest.Mock).mockImplementationOnce(() => {
+            throw error;
         });
 
         const isArtifactDownloaded = await downloadArtifact(repository, branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error extracting artifact TAR for ${repo}: ${errorMessage}`);
 
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        tar.x.mockReset();
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error extracting artifact TAR for ${repo}: ${error.message}`);
     });
 
     it('returns false if dependencies do not match', async () => {
@@ -818,8 +799,8 @@ describe('downloadArtifact', () => {
             },
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -833,22 +814,24 @@ describe('downloadArtifact', () => {
             },
         }));
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
         const existsSync = jest.spyOn(fs, 'existsSync');
-        existsSync.mockImplementation((path) => {
+        existsSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return true;
+            return false;
         });
 
         const readFileSync = jest.spyOn(fs, 'readFileSync');
-        readFileSync.mockImplementation((path) => {
+        readFileSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return JSON.stringify({
                 [dependency1]: dependency1Sha,
                 [dependency2]: dependency2OldSha,
             });
+            return '';
         });
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
@@ -859,12 +842,6 @@ describe('downloadArtifact', () => {
         expect(isArtifactDownloaded).toBe(false);
         expect(core.info).toHaveBeenCalledWith(`==> Found ${dependenciesPath}`);
         expect(core.warning).toHaveBeenCalledWith(`Error matching dependency ${dependency2} for ${repo}: ${dependency2NewSha} !== ${dependency2OldSha}`);
-
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        existsSync.mockReset();
-        readFileSync.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('extends environment object with install paths and dependency', async () => {
@@ -889,8 +866,8 @@ describe('downloadArtifact', () => {
             },
         };
 
-        Octokit.prototype.constructor.mockImplementation(() => ({
-            request: (route) => {
+        (Octokit.prototype.constructor as jest.Mock).mockImplementation(() => ({
+            request: (route: string) => {
                 switch (route) {
                 case 'GET /repos/{owner}/{repo}/actions/workflows/{workflow_id}/runs':
                     return resolveWorkflowRuns();
@@ -904,21 +881,17 @@ describe('downloadArtifact', () => {
             },
         }));
 
-        AdmZip.prototype.constructor.mockImplementation(() => ({
+        (AdmZip.prototype.constructor as jest.Mock).mockImplementationOnce(() => ({
             getEntries,
             extractAllTo,
         }));
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => true);
+        unlinkSync.mockImplementationOnce(() => true);
 
         const isArtifactDownloaded = await downloadArtifact(repository, branch, githubToken, downloadDir, installDir, os, compiler, testEnv);
 
         expect(isArtifactDownloaded).toBe(true);
         expect(testEnv).toStrictEqual(expectedEnv);
-
-        Octokit.prototype.constructor.mockReset();
-        AdmZip.prototype.constructor.mockReset();
-        unlinkSync.mockReset();
     });
 });
