@@ -1,10 +1,10 @@
-const core = require('@actions/core');
-const main = require('../src/main');
+import * as core from '@actions/core';
+import main from '../src/main';
 
 jest.mock('@actions/core');
 jest.mock('../src/main');
 
-const outputs = {
+const outputs: ActionOutputs = {
     bin_path: '/path/to/install/repo2/bin:/path/to/install/repo2/bin',
     include_path: '/path/to/install/repo2/include:/path/to/install/repo2/include',
     install_path: '/path/to/install/repo2:/path/to/install/repo2',
@@ -15,18 +15,16 @@ describe('entry', () => {
     it('sets path outputs and logs values', async () => {
         expect.assertions(8);
 
-        main.mockImplementation(() => Promise.resolve(outputs));
+        (main as jest.Mock).mockImplementationOnce(() => Promise.resolve(outputs));
 
         await jest.isolateModules(() => require('../src/index'));
 
         Object.keys(outputs).forEach((outputName) => {
-            const outputValue = outputs[outputName];
+            const outputValue = outputs[outputName as keyof ActionOutputs];
 
             expect(core.setOutput).toHaveBeenCalledWith(outputName, outputValue);
             expect(core.info).toHaveBeenCalledWith(`==> ${outputName}: ${outputValue}`);
         });
-
-        main.mockReset();
     });
 
     it('sets outputs with coverage file and logs values', async () => {
@@ -37,18 +35,16 @@ describe('entry', () => {
             coverage_file: '/path/to/repo/build/coverage.info',
         };
 
-        main.mockImplementation(() => Promise.resolve(outputsWithCoverageFile));
+        (main as jest.Mock).mockImplementationOnce(() => Promise.resolve(outputsWithCoverageFile));
 
         await jest.isolateModules(() => require('../src/index'));
 
         Object.keys(outputsWithCoverageFile).forEach((outputName) => {
-            const outputValue = outputsWithCoverageFile[outputName];
+            const outputValue = outputsWithCoverageFile[outputName as keyof ActionOutputs];
 
             expect(core.setOutput).toHaveBeenCalledWith(outputName, outputValue);
             expect(core.info).toHaveBeenCalledWith(`==> ${outputName}: ${outputValue}`);
         });
-
-        main.mockReset();
     });
 
     it('sets failure on errors', async () => {
@@ -56,17 +52,15 @@ describe('entry', () => {
 
         const errorMessage = 'Oops!';
 
-        main.mockImplementation(() => Promise.reject(errorMessage));
+        (main as jest.Mock).mockImplementationOnce(() => Promise.reject(errorMessage));
 
         // For some reason, checking toHaveBeenCalledWith() on this mock function does not work, possibly because of
         //   some race condition at play. Instead, we mock its implementation and check if it's called with correct
         //   parameter.
-        core.setFailed.mockImplementation((failureMessage) => {
+        (core.setFailed as jest.Mock).mockImplementation((failureMessage) => {
             expect(failureMessage).toBe(errorMessage);
         });
 
         await jest.isolateModules(() => require('../src/index'));
-
-        main.mockReset();
     });
 });
