@@ -1,11 +1,11 @@
-const fs = require('fs');
-const path = require('path');
-const core = require('@actions/core');
-const artifact = require('@actions/artifact');
-const filesize = require('filesize');
-const tar = require('tar');
+import fs from 'fs';
+import path from 'path';
+import * as core from '@actions/core';
+import * as artifact from '@actions/artifact';
+import filesize from 'filesize';
+import tar from 'tar';
 
-const uploadArtifact = require('../src/upload-artifact');
+import uploadArtifact from '../src/upload-artifact';
 
 jest.mock('@actions/core');
 jest.mock('@actions/artifact');
@@ -25,7 +25,8 @@ const rootDirectory = path.dirname(installDir);
 const tarPath = path.join(rootDirectory, tarName);
 const dependenciesName = `${artifactName}-dependencies.json`;
 const dependenciesPath = path.join(rootDirectory, dependenciesName);
-const errorMessage = 'Oops!';
+const errorObject = new Error('Oops!');
+const emptyObject = {};
 
 const dependencies = {
     'owner/repo1': 'de9f2c7fd25e1b3afad3e85a0bd17d9b100db4b3',
@@ -55,22 +56,22 @@ describe('uploadArtifact', () => {
             ...env,
         };
 
-        artifact.create.mockImplementation(() => ({
+        (artifact.create as jest.Mock).mockImplementationOnce(() => ({
             uploadArtifact: uploadResult,
         }));
 
         const statSync = jest.spyOn(fs, 'statSync');
-        statSync.mockImplementation(() => ({
+        (statSync as jest.Mock).mockImplementationOnce(() => ({
             size,
         }));
 
         const writeFileSync = jest.spyOn(fs, 'writeFileSync');
-        writeFileSync.mockImplementation((path) => {
+        writeFileSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return true;
         });
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => {
+        unlinkSync.mockImplementationOnce(() => {
             return true;
         });
 
@@ -80,11 +81,6 @@ describe('uploadArtifact', () => {
         expect(core.info).toHaveBeenCalledWith(`==> Created artifact TAR: ${tarPath} (${filesize(size)})`);
         expect(core.info).toHaveBeenCalledWith(`==> Created dependencies file: ${dependenciesPath}`);
         expect(core.info).toHaveBeenCalledWith(`==> Uploaded artifact: ${artifactName} (${filesize(size)})`);
-
-        artifact.create.mockReset();
-        statSync.mockReset();
-        writeFileSync.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('supports invalid repository name', async () => {
@@ -96,7 +92,7 @@ describe('uploadArtifact', () => {
 
         const coverageArtifactName = `coverage-${repo}-${os}-${compiler}`;
 
-        artifact.create.mockImplementation(() => ({
+        (artifact.create as jest.Mock).mockImplementationOnce(() => ({
             uploadArtifact: () => Promise.resolve({
                 artifactName: coverageArtifactName,
                 size,
@@ -105,12 +101,12 @@ describe('uploadArtifact', () => {
         }));
 
         const statSync = jest.spyOn(fs, 'statSync');
-        statSync.mockImplementation(() => ({
+        (statSync as jest.Mock).mockImplementationOnce(() => ({
             size,
         }));
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => {
+        unlinkSync.mockImplementationOnce(() => {
             return true;
         });
 
@@ -118,10 +114,6 @@ describe('uploadArtifact', () => {
 
         expect(isUploaded).toBe(true);
         expect(core.info).toHaveBeenCalledWith(`==> Uploaded artifact: ${coverageArtifactName} (${filesize(size)})`);
-
-        artifact.create.mockReset();
-        statSync.mockReset();
-        unlinkSync.mockReset();
     });
 
     it('constructs a different artifact name in case of ecbuild', async () => {
@@ -133,7 +125,7 @@ describe('uploadArtifact', () => {
 
         const ecbuildArtifactName = `ecbuild-${os}-cmake-${testEnv.CMAKE_VERSION}-${sha}`;
 
-        artifact.create.mockImplementation(() => ({
+        (artifact.create as jest.Mock).mockImplementationOnce(() => ({
             uploadArtifact: () => Promise.resolve({
                 artifactName: ecbuildArtifactName,
                 size,
@@ -142,17 +134,17 @@ describe('uploadArtifact', () => {
         }));
 
         const statSync = jest.spyOn(fs, 'statSync');
-        statSync.mockImplementation(() => ({
+        (statSync as jest.Mock).mockImplementationOnce(() => ({
             size,
         }));
 
         const writeFileSync = jest.spyOn(fs, 'writeFileSync');
-        writeFileSync.mockImplementation((path) => {
+        writeFileSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return true;
         });
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => {
+        unlinkSync.mockImplementationOnce(() => {
             return true;
         });
 
@@ -160,30 +152,29 @@ describe('uploadArtifact', () => {
 
         expect(isUploaded).toBe(true);
         expect(core.info).toHaveBeenCalledWith(`==> Uploaded artifact: ${ecbuildArtifactName} (${filesize(size)})`);
-
-        artifact.create.mockReset();
-        statSync.mockReset();
-        writeFileSync.mockReset();
-        unlinkSync.mockReset();
     });
 
-    it('returns false if creating artifact TAR fails', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('returns false if creating artifact TAR fails ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
         const testEnv = {
             ...env,
         };
 
-        tar.c.mockImplementation(() => {
-            throw new Error(errorMessage);
+        (tar.c as jest.Mock).mockImplementationOnce(() => {
+            throw error;
         });
 
         const isUploaded = await uploadArtifact(repository, sha, installDir, dependencies, os, compiler, testEnv);
 
         expect(isUploaded).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error creating artifact TAR for ${repo}: ${errorMessage}`);
 
-        tar.c.mockReset();
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error creating artifact TAR for ${repo}: ${error.message}`);
     });
 
     it('returns false if determining archive size errors out', async () => {
@@ -193,12 +184,12 @@ describe('uploadArtifact', () => {
             ...env,
         };
 
-        artifact.create.mockImplementation(() => ({
+        (artifact.create as jest.Mock).mockImplementation(() => ({
             uploadArtifact: uploadResult,
         }));
 
         const statSync = jest.spyOn(fs, 'statSync');
-        statSync.mockImplementation(() => ({
+        (statSync as jest.Mock).mockImplementationOnce(() => ({
             size: 0,
         }));
 
@@ -206,46 +197,46 @@ describe('uploadArtifact', () => {
 
         expect(isUploaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`Error determining size of artifact TAR for ${repo}`);
-
-        artifact.create.mockReset();
-        statSync.mockReset();
     });
 
-    it('returns false if writing dependencies file errors out', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('returns false if writing dependencies file errors out ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
         const testEnv = {
             ...env,
         };
 
-        artifact.create.mockImplementation(() => ({
+        (artifact.create as jest.Mock).mockImplementationOnce(() => ({
             uploadArtifact: uploadResult,
         }));
 
         const statSync = jest.spyOn(fs, 'statSync');
-        statSync.mockImplementation(() => ({
+        (statSync as jest.Mock).mockImplementationOnce(() => ({
             size,
         }));
 
         const writeFileSync = jest.spyOn(fs, 'writeFileSync');
-        writeFileSync.mockImplementation((path) => {
-            if (path === dependenciesPath) throw Error(errorMessage);
+        writeFileSync.mockImplementationOnce((path) => {
+            if (path === dependenciesPath) throw error;
         });
 
         const unlinkSync = jest.spyOn(fs, 'unlinkSync');
-        unlinkSync.mockImplementation(() => {
+        unlinkSync.mockImplementationOnce(() => {
             return true;
         });
 
         const isUploaded = await uploadArtifact(repository, sha, installDir, dependencies, os, compiler, testEnv);
 
         expect(isUploaded).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error writing dependencies file for ${repo}: ${errorMessage}`);
 
-        artifact.create.mockReset();
-        statSync.mockReset();
-        writeFileSync.mockReset();
-        unlinkSync.mockReset();
+        (artifact.create as jest.Mock).mockReset();
+
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error writing dependencies file for ${repo}: ${error.message}`);
     });
 
     it('returns false if artifact item upload has some failures', async () => {
@@ -255,7 +246,7 @@ describe('uploadArtifact', () => {
             ...env,
         };
 
-        artifact.create.mockImplementation(() => ({
+        (artifact.create as jest.Mock).mockImplementationOnce(() => ({
             uploadArtifact: () => Promise.resolve({
                 artifactName,
                 size,
@@ -266,12 +257,12 @@ describe('uploadArtifact', () => {
         }));
 
         const statSync = jest.spyOn(fs, 'statSync');
-        statSync.mockImplementation(() => ({
+        (statSync as jest.Mock).mockImplementationOnce(() => ({
             size,
         }));
 
         const writeFileSync = jest.spyOn(fs, 'writeFileSync');
-        writeFileSync.mockImplementation((path) => {
+        writeFileSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return true;
         });
 
@@ -279,10 +270,6 @@ describe('uploadArtifact', () => {
 
         expect(isUploaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`Error uploading artifact for ${repo}: ${artifactName}`);
-
-        artifact.create.mockReset();
-        statSync.mockReset();
-        writeFileSync.mockReset();
     });
 
     it('returns false if artifact item upload returns empty result', async () => {
@@ -292,17 +279,17 @@ describe('uploadArtifact', () => {
             ...env,
         };
 
-        artifact.create.mockImplementation(() => ({
+        (artifact.create as jest.Mock).mockImplementationOnce(() => ({
             uploadArtifact: () => Promise.resolve(),
         }));
 
         const statSync = jest.spyOn(fs, 'statSync');
-        statSync.mockImplementation(() => ({
+        (statSync as jest.Mock).mockImplementationOnce(() => ({
             size,
         }));
 
         const writeFileSync = jest.spyOn(fs, 'writeFileSync');
-        writeFileSync.mockImplementation((path) => {
+        writeFileSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return true;
         });
 
@@ -310,40 +297,38 @@ describe('uploadArtifact', () => {
 
         expect(isUploaded).toBe(false);
         expect(core.warning).toHaveBeenCalledWith(`Error uploading artifact for ${repo}`);
-
-        artifact.create.mockReset();
-        statSync.mockReset();
-        writeFileSync.mockReset();
     });
 
-    it('returns false if artifact item upload fails', async () => {
-        expect.assertions(2);
+    it.each`
+        error
+        ${errorObject}
+        ${emptyObject}
+    `('returns false if artifact item upload fails ($error)', async ({ error }) => {
+        expect.hasAssertions();
 
         const testEnv = {
             ...env,
         };
 
-        artifact.create.mockImplementation(() => ({
-            uploadArtifact: () => Promise.reject(new Error(errorMessage)),
+        (artifact.create as jest.Mock).mockImplementationOnce(() => ({
+            uploadArtifact: () => Promise.reject(error),
         }));
 
         const statSync = jest.spyOn(fs, 'statSync');
-        statSync.mockImplementation(() => ({
+        (statSync as jest.Mock).mockImplementationOnce(() => ({
             size,
         }));
 
         const writeFileSync = jest.spyOn(fs, 'writeFileSync');
-        writeFileSync.mockImplementation((path) => {
+        writeFileSync.mockImplementationOnce((path) => {
             if (path === dependenciesPath) return true;
         });
 
         const isUploaded = await uploadArtifact(repository, sha, installDir, dependencies, os, compiler, testEnv);
 
         expect(isUploaded).toBe(false);
-        expect(core.warning).toHaveBeenCalledWith(`Error uploading artifact for ${repo}: ${errorMessage}`);
 
-        artifact.create.mockReset();
-        statSync.mockReset();
-        writeFileSync.mockReset();
+        if (!(error instanceof Error)) return;
+        expect(core.warning).toHaveBeenCalledWith(`Error uploading artifact for ${repo}: ${error.message}`);
     });
 });
